fix: extracted SearchBar fun to separate file for multi-use
